// collabarative-design-tool-main/src/types.ts
export type Color = {
  r: number;
  g: number;
  b: number;
};

export type Camera = {
  x: number;
  y: number;
  zoom: number;
};

export enum LayerType {
  Rectangle = "Rectangle",
  Ellipse = "Ellipse",
  Path = "Path",
  Text = "Text",
}

export type RectangleLayer = {
  type: LayerType.Rectangle;
  x: number;
  y: number;
  width: number;
  height: number;
  fill: Color;
  stroke: Color;
  opacity: number;
  cornerRadius?: number;
};
export type EllipseLayer = {
  type: LayerType.Ellipse;
  x: number;
  y: number;
  width: number;
  height: number;
  fill: Color;
  stroke: Color;
  opacity: number;
  cornerRadius?: number;
};

export type PathLayer = {
  type: LayerType.Path;
  x: number;
  y: number;
  width: number;
  height: number;
  fill: Color;
  stroke: Color;
  opacity: number;
  points: number[][];
};

export type TextLayer = {
  type: LayerType.Text;
  x: number;
  y: number;
  height: number;
  width: number;
  text: string;
  fontSize: number;
  fontWeight: number;
  fontFamily: string;
  fill: Color;
  stroke: Color;
  opacity: number;
};

export type Layer = RectangleLayer | EllipseLayer | PathLayer | TextLayer;

export type Point = {
  x: number;
  y: number;
};

export type CanvasState =
  | {
      mode: CanvasMode.None;
    }
  | {
      mode: CanvasMode.Inserting;
      layerType: LayerType.Rectangle | LayerType.Ellipse | LayerType.Text;
    }
  |
  {
    mode:CanvasMode.Dragging;
    origin:Point|null;
  }
  | 
  {
    mode: CanvasMode.Pencil;
  }
  |
  {
    mode:CanvasMode.Resizing,
    initialBounds:XYWH
    corner:Side
  }
  |
  {
    mode:CanvasMode.Translating;
    current:Point;
  }
  ;

export enum CanvasMode {
  None,
  Dragging,
  Inserting,
  Pencil,
  Resizing,
  Translating
}

export type XYWH={
  x:number,
  y:number,
  width:number,
  height:number
}

export enum Side{
   Top=1,
   Bottom=2,
   Left=4,
   Right=8
}